feat: address comments from vishwa and move validation tests to its own folder
