+++ CMakeLists.txt
@@ -463,7 +463,7 @@ IF (BUILD_VALIDATION_TESTS)
 
     ENABLE_TESTING ()
 
-    FILE (GLOB_RECURSE TEST_SRCS "${PROJECT_SOURCE_DIR}/test/Validation/*.validation.cpp" "${PROJECT_SOURCE_DIR}/test/Validation/*.hpp" "${PROJECT_SOURCE_DIR}/test/Validation/*.cpp")
+    FILE (GLOB_RECURSE TEST_SRCS "${PROJECT_SOURCE_DIR}/validation/${PROJECT_PATH}/*.cpp")
 
     ADD_EXECUTABLE (${VALIDATION_TESTS_TARGET} ${TEST_SRCS})
 
@@ -484,7 +484,7 @@ IF (BUILD_VALIDATION_TESTS)
 
     TARGET_INCLUDE_DIRECTORIES (${VALIDATION_TESTS_TARGET} PUBLIC ${GTEST_INCLUDE_DIRS})
     TARGET_INCLUDE_DIRECTORIES (${VALIDATION_TESTS_TARGET} PUBLIC "${PROJECT_SOURCE_DIR}/include")
-    TARGET_INCLUDE_DIRECTORIES (${VALIDATION_TESTS_TARGET} PUBLIC "${PROJECT_SOURCE_DIR}/test")
+    TARGET_INCLUDE_DIRECTORIES (${VALIDATION_TESTS_TARGET} PUBLIC "${PROJECT_SOURCE_DIR}/validation")
 
     # TARGET_LINK_LIBRARIES (${VALIDATION_TESTS_TARGET} "GTest::GTest" "GTest::Main")
     TARGET_LINK_LIBRARIES (${VALIDATION_TESTS_TARGET} "${GTEST_BOTH_LIBRARIES}")
@@ -495,7 +495,7 @@ IF (BUILD_VALIDATION_TESTS)
 
     SET_TARGET_PROPERTIES (${VALIDATION_TESTS_TARGET} PROPERTIES VERSION ${PROJECT_VERSION_STRING} OUTPUT_NAME ${VALIDATION_TESTS_TARGET} CLEAN_DIRECT_OUTPUT 1 INSTALL_RPATH "$ORIGIN/../lib:$ORIGIN/")
 
-    INSTALL (TARGETS ${VALIDATION_TESTS_TARGET} DESTINATION ${INSTALL_TEST} COMPONENT "tests")
+    INSTALL (TARGETS ${VALIDATION_TESTS_TARGET} DESTINATION ${INSTALL_TEST} COMPONENT "validation")
 
 ENDIF ()
