+++ README.md
@@ -221,14 +221,103 @@ Benchmarks are pushed to the [GitHub pages](https://open-space-collective.github
 
 ### Validation
 
-OSTk has a cross-validation framework built into it, which allows us to compare its end-to-end propagation accuracy of a full "mission sequence" to other flight qualified tools like GMAT and Orekit.
+OSTk has a cross-validation framework built into it (inside the `/validation` folder), which allows us to compare its end-to-end propagation accuracy of a full "mission sequence" to other flight qualified tools like GMAT and Orekit.
 
-A standardized input format (a yaml file) to define the particular mission sequence scenario is used, and is tooling agnostic.
+#### Scenario Definition
 
-The validation framework takes this yaml scenario definition (inside the /test/Validation/data/scenarios folder) and runs the scenario using OSTk. Next, it compares the outputs (things like the position, velocity, mass, acceleration of the spacecraft) at each timestep during the scenario and checks if they are within a certain tolerance of those same quantities generated by GMAT or Orekit in another repo and read in as a .csv file.
+A standardized input format (a yaml file) to define the particular mission sequence scenario is used, and is tooling agnostic. An example of it is shown below.
 
-Each scenario, the tool it is being compared against, what output quantities are being compared and to what tolerance are defined in the `parameters` variable inside the `Framework.validation.cpp` file, and can be modified or have new scenarios and tolerances appended to them.
+- The spacecraft's parameters and initial state are defined under the `spacecraft` header
+- The sequence header contains the dynamic forces to be applied during propagation, the type of numerical propagator to use, and the segments (burning or coasting) to execute during the mission sequence
+- Finally, the output header contains the quantities that we want to output (in this case time elapsed, position, and velocity), and the interval at which to report them
 
+```yaml
+type: MISSION_SEQUENCE
+data:
+  spacecraft:
+    mass: 100.0
+    drag-cross-section: 1.0
+    drag-coefficient: 2.2
+    orbit:
+      type: CARTESIAN
+      data:
+        date:
+          time-scale: UTC
+          value: "2023-01-01T00:00:00"
+        frame: GCRF
+        body: EARTH
+        x: -4283387.412456233
+        y: -4451426.776125101
+        z: -2967617.850750065
+        vx: 4948.074939732174
+        vy: -957.3429532772124
+        vz: -5721.173027553034
+
+  sequence:
+    max-duration: 86400.1
+    forces:
+      - type: GRAVITY
+        data:
+          body: EARTH
+          model: EGM96
+          degree: 0
+          order: 0
+    propagator:
+      type: RUNGE_KUTTA_DORMAND_PRINCE_45
+      data:
+        initial-step: 30.0
+        min-step: 0.001
+        max-step: 2700.0
+        relative-tolerance: 1.0e-12
+        absolute-tolerance: 1.0e-12
+    segments:
+      - type: COAST
+        data:
+          stop-condition:
+            type: RELATIVE_TIME
+            data:
+              duration: 86400.0
+
+  output:
+    step: 120.0
+    include:
+      - ELAPSED_SECONDS
+      - CARTESIAN_POSITION_GCRF
+      - CARTESIAN_VELOCITY_GCRF
+```
+
+#### Cross Validation and Accuracy Tolerances
+
+The validation framework takes this yaml scenario definition (from the `/validation/data/scenarios` folder) and runs the scenario using OSTk. Next, it compares the specified outputs (quantities like the position, velocity, mass, acceleration of the spacecraft) at the desired reporting step generated by OSTk with each external validation tool (whose outputs were generated via the same process and are in the `/validation/data/gmat_astrodynamics` or `/validation/data/orekit_astrodynamics` folder as `.csv`), to ensure they are within a certain tolerance of each other.
+
+**To add a scenario:**
+1. Write a scenario definition using the `.yaml` format above, and name it `scenarioX-mission-sequence.yaml`, and put it in wth the others. If the format isn't correct or there is another issue, OSTk will throw an error when validating.
+2. Generate the `.csv` outputs from the external tools you want to compare OSTk against (repos with an GMAT and Orekit wrapper that can ingest this `.yaml` format do exist and will be open sourced soon).
+3. Add another test case to the `Framework.validation.cpp` file, with the scenario name, the tool to compare against, and the output quantities and comparison tolerances for those. An exampe of this is provided below.
+
+```cpp
+{
+    "scenario001-mission-sequence",  // Spherical gravity only
+    {
+        {
+            Tool::GMAT,
+            {
+                {Quantity::CARTESIAN_POSITION_GCRF, 1.1e-0},
+                {Quantity::CARTESIAN_VELOCITY_GCRF, 1.2e-3},
+            },
+        },
+        {
+            Tool::OREKIT,
+            {
+                {Quantity::CARTESIAN_POSITION_GCRF, 1.1e-0},
+                {Quantity::CARTESIAN_VELOCITY_GCRF, 1.2e-3},
+            },
+        },
+    },
+},
+```
+
+#### Running the Validation Tests
 The validation tests can be run with `ostk-validate` from within the dev container, or `make validation` as a standalone.
 
 ## Dependencies
