--- test/Validation/MissionSequence.hpp
@@ -1,210 +1,0 @@
-/// Apache License 2.0
-
-#ifndef __OpenSpaceToolkit_Astrodynamics_Validation_MissionSequence__
-#define __OpenSpaceToolkit_Astrodynamics_Validation_MissionSequence__
-
-#include <Validation/CrossValidator.hpp>
-#include <Validation/Parser.hpp>
-
-#include <OpenSpaceToolkit/Core/Containers/Array.hpp>
-#include <OpenSpaceToolkit/Core/Containers/Dictionary.hpp>
-#include <OpenSpaceToolkit/Core/Containers/Table.hpp>
-#include <OpenSpaceToolkit/Core/Containers/Tuple.hpp>
-#include <OpenSpaceToolkit/Core/Types/Integer.hpp>
-#include <OpenSpaceToolkit/Core/Types/Real.hpp>
-#include <OpenSpaceToolkit/Core/Types/Shared.hpp>
-#include <OpenSpaceToolkit/Core/Types/Size.hpp>
-#include <OpenSpaceToolkit/Core/Types/String.hpp>
-
-#include <OpenSpaceToolkit/Mathematics/Geometry/3D/Objects/Composite.hpp>
-#include <OpenSpaceToolkit/Mathematics/Geometry/3D/Objects/Cuboid.hpp>
-#include <OpenSpaceToolkit/Mathematics/Geometry/3D/Objects/Point.hpp>
-#include <OpenSpaceToolkit/Mathematics/Objects/Vector.hpp>
-
-#include <OpenSpaceToolkit/Physics/Data/Scalar.hpp>
-#include <OpenSpaceToolkit/Physics/Environment.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Atmospheric/Earth.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Atmospheric/Earth/CSSISpaceWeather.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Atmospheric/Earth/Exponential.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Atmospheric/Earth/Manager.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Ephemerides/Analytical.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Gravitational/Earth.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Magnetic/Earth.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Object.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Objects/CelestialBodies/Earth.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Objects/CelestialBodies/Moon.hpp>
-#include <OpenSpaceToolkit/Physics/Environment/Objects/CelestialBodies/Sun.hpp>
-#include <OpenSpaceToolkit/Physics/Time/DateTime.hpp>
-#include <OpenSpaceToolkit/Physics/Time/Duration.hpp>
-#include <OpenSpaceToolkit/Physics/Time/Instant.hpp>
-#include <OpenSpaceToolkit/Physics/Time/Interval.hpp>
-#include <OpenSpaceToolkit/Physics/Time/Scale.hpp>
-#include <OpenSpaceToolkit/Physics/Time/Time.hpp>
-#include <OpenSpaceToolkit/Physics/Units/Derived.hpp>
-#include <OpenSpaceToolkit/Physics/Units/Length.hpp>
-#include <OpenSpaceToolkit/Physics/Units/Mass.hpp>
-
-#include <OpenSpaceToolkit/Astrodynamics/Dynamics/AtmosphericDrag.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Dynamics/CentralBodyGravity.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Dynamics/PositionDerivative.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Dynamics/ThirdBodyGravity.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Dynamics/Thruster.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/EventCondition/InstantCondition.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/GuidanceLaw/ConstantThrust.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/GuidanceLaw/QLaw.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Solvers/FiniteDifferenceSolver.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/LocalOrbitalFrameDirection.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/Orbit.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/Orbit/Model.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/Orbit/Models/Kepler/COE.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/Propagator.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/Segment.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/Sequence.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesBroker.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesSubset.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesSubsets/CartesianPosition.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesSubsets/CartesianVelocity.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/NumericalSolver.hpp>
-#include <OpenSpaceToolkit/Astrodynamics/Trajectory/StateBuilder.hpp>
-
-namespace ostk
-{
-namespace astro
-{
-namespace validation
-{
-
-using ostk::core::ctnr::Array;
-using ostk::core::ctnr::Table;
-using ostk::core::ctnr::Tuple;
-using CntrObject = ostk::core::ctnr::Object;
-using ostk::core::ctnr::Dictionary;
-using ostk::core::filesystem::File;
-using ostk::core::filesystem::Path;
-using ostk::core::types::Integer;
-using ostk::core::types::Real;
-using ostk::core::types::Shared;
-using ostk::core::types::String;
-using ostk::core::types::Size;
-
-using ostk::math::geometry::d3::objects::Composite;
-using ostk::math::geometry::d3::objects::Cuboid;
-using ostk::math::geometry::d3::objects::Point;
-using ostk::math::object::Matrix3d;
-using ostk::math::object::Vector3d;
-using ostk::math::object::VectorXd;
-using ostk::math::object::MatrixXd;
-
-using ostk::physics::data::Scalar;
-using ostk::physics::Environment;
-using ostk::physics::coord::Frame;
-using ostk::physics::coord::Position;
-using ostk::physics::coord::Velocity;
-using ostk::physics::environment::Object;
-using ostk::physics::environment::object::Celestial;
-using ostk::physics::environment::object::celestial::Earth;
-using ostk::physics::environment::object::celestial::Moon;
-using ostk::physics::environment::object::celestial::Sun;
-using ostk::physics::environment::ephemerides::Analytical;
-using ostk::physics::time::DateTime;
-using ostk::physics::time::Duration;
-using ostk::physics::time::Instant;
-using ostk::physics::time::Interval;
-using ostk::physics::time::Time;
-using ostk::physics::time::Scale;
-using ostk::physics::units::Angle;
-using ostk::physics::units::Derived;
-using ostk::physics::units::Length;
-using ostk::physics::units::Mass;
-using ostk::physics::environment::object::celestial::Earth;
-using EarthGravitationalModel = ostk::physics::environment::gravitational::Earth;
-using EarthMagneticModel = ostk::physics::environment::magnetic::Earth;
-using EarthAtmosphericModel = ostk::physics::environment::atmospheric::Earth;
-using EarthSpaceWeatherManager = ostk::physics::environment::atmospheric::earth::Manager;
-using ostk::physics::environment::atmospheric::earth::CSSISpaceWeather;
-
-using ostk::astro::eventcondition::InstantCondition;
-using ostk::astro::trajectory::State;
-using ostk::astro::trajectory::StateBuilder;
-using ostk::astro::trajectory::Sequence;
-using ostk::astro::trajectory::Segment;
-using ostk::astro::trajectory::Propagator;
-using ostk::astro::trajectory::Orbit;
-using ostk::astro::trajectory::LocalOrbitalFrameFactory;
-using ostk::astro::trajectory::LocalOrbitalFrameDirection;
-using ostk::astro::trajectory::state::NumericalSolver;
-using ostk::astro::trajectory::state::CoordinatesSubset;
-using ostk::astro::trajectory::state::CoordinatesBroker;
-using ostk::astro::trajectory::state::coordinatessubsets::CartesianPosition;
-using ostk::astro::trajectory::state::coordinatessubsets::CartesianVelocity;
-using ostk::astro::trajectory::state::NumericalSolver;
-using ostk::astro::trajectory::orbit::models::kepler::COE;
-using ostk::astro::Dynamics;
-using ostk::astro::flight::system::PropulsionSystem;
-using ostk::astro::flight::system::SatelliteSystem;
-using ostk::astro::dynamics::PositionDerivative;
-using ostk::astro::dynamics::CentralBodyGravity;
-using ostk::astro::dynamics::ThirdBodyGravity;
-using ostk::astro::dynamics::AtmosphericDrag;
-using ostk::astro::dynamics::Thruster;
-using ostk::astro::guidancelaw::ConstantThrust;
-using ostk::astro::guidancelaw::QLaw;
-using ostk::astro::solvers::FiniteDifferenceSolver;
-
-using ostk::astro::validation::Parser;
-using ostk::astro::validation::Quantity;
-using ostk::astro::validation::Tool;
-using ostk::astro::validation::ToolComparison;
-using ostk::astro::validation::QuantityComparison;
-
-/// @brief Holds the OSTk objects and data required to define and run a "Mission Sequence".
-class MissionSequence
-{
-   public:
-    /// @brief Constructor
-    /// @param aDataTree A dictionary containing all the data needed to define and initialize a "Mission Sequence".
-    MissionSequence(const Dictionary& aDataTree);
-
-    /// @brief Destructor
-    ~MissionSequence();
-
-    /// @brief Run the "Mission Sequence".
-    void run();
-
-    /// @brief Access the initial state and satellite system.
-    const Pair<State, SatelliteSystem>& accessInitialStateAndSatelliteSystem() const;
-
-    /// @brief Access the environment.
-    const Environment& accessEnvironment() const;
-
-    /// @brief Access the dynamics.
-    const Array<Shared<Dynamics>>& accessDynamics() const;
-
-    /// @brief Access the sequence.
-    const Sequence& accessSequence() const;
-
-    /// @brief Access the solved states.
-    const Array<State>& accessSolvedStates() const;
-
-    /// @brief Compare the results of the "Mission Sequence" with the results of a tool.
-    /// @param aReferenceData A table containing the reference data from the tool.
-    /// @param aToolComparison A tool comparison object containing the tool name and the quantities to compare.
-    /// @return An array of vectors containing the differences between the results of the "Mission Sequence" and the
-    /// results of the tool.
-    Array<VectorXd> compareResults(const Table& referenceData, const ToolComparison& aToolComparison) const;
-
-   private:
-    Dictionary dataTree_;
-    Pair<State, SatelliteSystem> initialStateAndSatelliteSystem_;
-    Environment environment_;
-    Array<Shared<Dynamics>> dynamics_;
-    Sequence sequence_;
-    Array<State> solvedStates_;
-};
-
-}  // namespace validation
-}  // namespace astro
-}  // namespace ostk
-
-#endif
