+++ validation/OpenSpaceToolkit/Astrodynamics/CrossValidator.cpp
@@ -1,6 +1,6 @@
 /// Apache License 2.0
 
-#include <Validation/CrossValidator.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/CrossValidator.hpp>
 
 namespace ostk
 {
