+++ validation/OpenSpaceToolkit/Astrodynamics/data/gmat_astrodynamics/scenario002-mission-sequence.csv
