+++ validation/OpenSpaceToolkit/Astrodynamics/data/scenarios/scenario001-mission-sequence.yaml
