+++ validation/OpenSpaceToolkit/Astrodynamics/data/scenarios/scenario002-mission-sequence.yaml
