+++ validation/OpenSpaceToolkit/Astrodynamics/data/scenarios/scenario003-mission-sequence.yaml
