+++ validation/OpenSpaceToolkit/Astrodynamics/data/scenarios/scenario004-mission-sequence.yaml
