+++ validation/OpenSpaceToolkit/Astrodynamics/data/scenarios/scenario005-mission-sequence.yaml
